+++ types/address.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/tendermint/tendermint/crypto"
-	"github.com/tendermint/tendermint/crypto/encoding/amino"
+	cryptoAmino "github.com/tendermint/tendermint/crypto/encoding/amino"
 	"github.com/tendermint/tendermint/libs/bech32"
 )
 
@@ -19,6 +19,8 @@ const (
 	AddrLen = 20
 	// VoteAddrLen defines a valid vote address length
 	VoteAddrLen = 48
+	// BLSSignatureLength defines the length of a BLS signature
+	BLSSignatureLength = 96
 	// SmartChainAddressLength defines a valid smart chain address length
 	SmartChainAddressLength = 20